Update parse_phenotype.py
README: I have edited the script to include the columns that need parsing (I had to do it in this way because all the names are different between baseline and oneweek). I encountered an issue with the columns that I have left unquoted on lines 41 43 and 45. These column names also appear in other columns and I dont know how to use the regex comands to seperate these. Could you provide some help on this?